build: :wrench: add Quarto build config files (#29)
Adds:

- `_brand.yml` file that contains the website theming
- `_quarto.yml` that contains Quarto website generation settings.
- `_publish.yml` that contains the settings for connecting to Netlify.
These need to be updated when @danielibsen can setup the repo on
Netlify.
- `index.qmd` that is empty, but needed in order for the website to
build.

I haven't added any other docs to the navbar, as that will be step PRs
and at a later point when we decide what folders/files should be called
and what should be added to the website.