feat: :sparkles: storage and backup section (#86)
# Description

I'm not sure if these should be released as PRs one at a time, as there
are probably things that look like they might be missing but may be
documented in a different part of the overall plan. We'll just have to
see if this makes sense to review without the other sections.

This part is focused on the physical storage of data, as well as access
to raw data, not access to data which will be used for analysis. It will
also not touch on Seedcase Sprout, this will be part of Data Collection
as also Responsibilities & Resources.

Closes #81 

This PR needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: Daniel Ibsen <48243937+danielibsen@users.noreply.github.com>