feat: :sparkles: naming convention for files and variables (#146)
# Description

I've tried to put together the guidance on naming conventions for files
and variables based mostly on the established practices from SDCC. I've
placed it in a new folder. We can have a chat about whether or not we
want to move it to a different repo later.

Closes #132 

This PR needs an in-depth review.

## Checklist

- [x] Formatted Markdown
- [x] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>